ci: compile under a consumer's warning flags, not only under ours (#86)
METL is HEADER-ONLY, which means its headers are compiled with the
CONSUMER's flags and not with METL's. The project builds its own tests under
-Wall -Wextra -Wpedantic -Wshadow and friends; nothing checked what a caller
sees under the stricter sets embedded work actually uses. -Wconversion is
close to standard practice in MISRA-adjacent shops.

CMake does not shield them either. CMakeLists.txt exports the include
directory WITHOUT `SYSTEM`, so even find_package(metl) consumers compile
these headers under their own -Werror. Verified, not assumed:

    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>   # no SYSTEM
    $<INSTALL_INTERFACE:include>

Measuring it found four real diagnostics, one of them in a public header:

  include/metl/bit.hpp:237            `T{1} << width` promotes the operand to
                                      int and narrows on return. Correct --
                                      the width check bounds it -- but it
                                      lands in every consumer's build log.
  crc_table_equivalence_test.cpp x2   `crc ^= byte` and `crc << 1u` promote
                                      and convert back.
  spsc_queue_cached_index_test.cpp    int to size_t.

And a fifth that is not a defect: fuzz_vocab.cpp writes `v = v` on purpose,
to exercise the self-assignment path where "destroy the old, construct the
new" destroys the thing it is about to read. Suppressed at that line and
nowhere wider, because a file- or build-level suppression would also hide an
accidental self-assignment.

The flag list is the claim, and it is broader than the four fixes needed:
thirteen flags, all of them already free, checked across 116 translation
units under clang and gcc, C++17 and C++20. Every one measured at zero
before being added -- the job locks in what is true rather than announcing
an aspiration.

The TU-count guard is there for the same reason bench-smoke needed one: a
glob that stops matching turns a loop into a no-op that passes forever. The
two excluded directories are excluded in the open -- tests/embedded is
cross-only, and tests/compile_fail is SUPPOSED to fail.

Found while scoping this job: metl::visit silently truncating a visitor's
result (#84). The -Wconversion diagnostic at variant.hpp:673 was the symptom;
that PR fixed the cause and removed the warning, which is why it is not in
the list above.

Verified: 116 TUs clean under the exact flag list on clang C++17 and C++20,
89/89 ctest, 8/8 compile-failure cases, check_docs and clang-format clean.
gcc is the one combination that cannot be checked here -- its -Wconversion is
more aggressive than clang's -- so CI is the first run of those two rows.

+++ tests/sync/spsc_queue_cached_index_test.cpp
@@ -94,7 +94,7 @@ int main() {
     std::uint32_t next_pop = 0;
 
     for (int round = 0; round < 16; ++round) {
-      const std::size_t burst = (round % 3) + 1;  // 1, 2, 3, 1, ...
+      const std::size_t burst = static_cast<std::size_t>((round % 3) + 1);  // 1, 2, 3, 1, ...
       for (std::size_t i = 0; i < burst; ++i) {
         CHECK(queue.try_push(next_push++));
       }
